MICS-11962 Keycloak createLoginUrl fix
When going to navigator, there was an error on the console:
Cannot get version: TypeError: Cannot read properties of undefined
(reading 'authorize') at r.s.createLoginUrl

This error occurs in getVersionLoop() in the Version saga of
navigator. It calls navigatorService.getVersion which calls the
ApiService.getRequest with localUrl set to true.

When the ApiService makes the call, if using Keycloak, it tries
to embed the call within the KeycloakService.updateToken.

This caused the error in our case.
Therefore, this commit does not allow to update the Keycloak token
when localUrl is set to true in the call, in order to avoid the
error.