MICS-13153 Fix charts bottom padding problem
In the editable mode we add the top edit panel.
A side effect is that the bottom side of charts
becomes hidden as the height of the card is the same.
We cannot easily modify the size of the card, as
it is defined in units and calculated inside ReactGridLayout.
So in this fix we slightly reduce the height of the charts
to fit the card in editable mode.
Fixed some parameters for horizontal cards height calculation.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>