MICS-14185 Add charts defined by id
We want to use charts defined by it's id in dashboard editor:
{"chart_id":"XX"}
The main part for the visual editor is implemented in adv.
But on the application side on the tab "Advanced" we have the
json text editor, that should be also functional for editing
such charts.
So we hide externally loaded properties when we fill the text
editor, and restore them when refill visual components.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>