MICS-14185 Fix problem of dashboard editor tab switch
When we modify the json code, the modification is launched
with a second delay.
If we switch fastly to the wysiwyg tab, the page is crushed.
The reason is that the react context/state is deprecated for
the scheduled code.
In this commit we cancel the last lauched timer, by the start
of a new actualized timer.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>