MICS-12839 Add ability to format non-date x values for area
The component AreaChart accepts xKey parameter as a string or
as a XKey object. Here we added the ability to show the X values
as a string (Category), without formatting they as dates.
For that, the component checks if the X values are all valid dates.
If not, the component don't use date formatting and switch the
underlying Highchart component to "Category" mode.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>