MICS-14305 use charts panel to edit dashboard charts
* add modification and save chart functions

* chart search panel now can take selected item parameter
also as a string (id)

* remove external charts loading logic from DashboardLayout
We don't need it anymore, as charts are loaded by the
search panel component (at least for now)

* we also don't need to remove externally loaded properties
from the chart config, when we open it in editor. On the
contrary, the editor now shows exactly externally loaded
config in the JSON editor without chart_id.

* now we doing some refresh of our components (charts list
and json editor) when we switch between tabs, because the
hidden tab is not correctly re-rendered. So, the editor is
rendered as a simple empty block when it is on the inactive
tab and re-rendered when it's tab becomes active. And for
the list, we reset the selected chart (if it's json code
were modified) when the tab becomes active.

* open the JSON editor tab by default for editing
json-defined charts, and charts list for other cases

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>