MICS-14889 Fix save series title for saved charts in query tools
Here we fix the case where we have Series with dimentions.
When we save such query, the title of the series is not saved.
Even when we save it, it is not loaded.
So we fix the save and the load by passing the name between promises.
Also, we add a temporary type to formalize the list of it's properties.

Signed-off-by: Igor Epimakhov <iepimakhov@mediarithmics.com>